Update GenericNode for field management and add related tests
The GenericNode code has been enhanced to better manage fields, adding a setFields method that creates a new HashMap, as well as a getField method that retrieves field values based on a given name. This commit also includes the introduction of a new test, GenericNodeTest, which verifies the proper functioning of these new field management methods. The import statements have also been reorganized in various classes for better code readability.