Update data handling in ValidationIssue and associated files
Changed the data type in ValidationIssue.Context from JsonNode to Object, to allow handling data as a tree instead of a string. This refactoring also involved updating all the related methods and constructors. The associated files like JsonUtil, and test classes have been adjusted accordingly to fit this new strategy. This change simplifies the handling of JSON data.