Add dimension resolution and range splitting functionality
The commit introduces methods to resolve a dimension index and split a range into sub-ranges in the classes ZPoint and ZRange. Additionally, it provides utility mechanisms to create an IPF Shard in the IPFSignatureAgreementConstraintTest. This revision enhances internal classes by enriching their functionalities and paves the way for further possible optimizations and utilizations. The unit tests for the newly introduced methods have been updated accordingly.