Add IntStream method and refactor methods in ZTensor
An IntStream method called valueStream was introduced in the ZTensor class to provide a stream of the tensor's values, which can be used by various methods. The allMatch and anyMatch methods have been refactored to use the new valueStream method, resulting in cleaner and more concise code. The rearrangement of some methods has also been done for a better logical flow.