Reorder plugins in pom.xml
The Apache Maven plugins in the pom.xml file have been reorganized. The 'spotless-maven-plugin' has been moved up, while the 'maven-enforcer-plugin' and 'maven-compiler-plugin' have been reordered to the latter part of the configuration list, along with their associated settings. This reordering aims to improve the organization of the pom file and make the plugin usage clearer.